refactor: move receiveMessages to sync-receive
